Converted many long properties to ints
Converted many long properties to ints.  Left some as longs, particularly those related to money.

+++ src/ArgentPonyWarcraftClient/Models/GameDataApi/Achievement/Achievement.cs
@@ -17,7 +17,7 @@ namespace ArgentPonyWarcraftClient
         /// Gets the ID of this achievement.
         /// </summary>
         [JsonPropertyName("id")]
-        public long Id { get; private set; }
+        public int Id { get; private set; }
 
         /// <summary>
         /// Gets a reference to the achievement category to which this achievement belongs.
