+++ src/ArgentPonyWarcraftClient/Models/GameDataApi/Profession/SkillTier.cs
@@ -17,7 +17,7 @@ namespace ArgentPonyWarcraftClient
         /// Gets the ID of the skill tier.
         /// </summary>
         [JsonPropertyName("id")]
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         /// <summary>
         /// Gets the name of the skill tier.
@@ -29,13 +29,13 @@ namespace ArgentPonyWarcraftClient
         /// Gets the minimum skill level for the skill tier.
         /// </summary>
         [JsonPropertyName("minimum_skill_level")]
-        public long MinimumSkillLevel { get; set; }
+        public int MinimumSkillLevel { get; set; }
 
         /// <summary>
         /// Gets the maximum skill level for the skill tier.
         /// </summary>
         [JsonPropertyName("maximum_skill_level")]
-        public long MaximumSkillLevel { get; set; }
+        public int MaximumSkillLevel { get; set; }
 
         /// <summary>
         /// Gets the categories in the skill tier.
